struct pack for gigadrip

+++ test/mocks/MockGIGADRIP20.sol
@@ -7,18 +7,18 @@ contract MockGIGADRIP20 is GIGADRIP20 {
         string memory _name,
         string memory _symbol,
         uint8 _decimals,
-        uint256 _emissionRatePerBlock
+        uint128 _emissionRatePerBlock
     ) GIGADRIP20(_name, _symbol, _decimals, _emissionRatePerBlock) {}
 
     function mint(uint256 amount) external {
         _mint(msg.sender, amount);
     }
 
-    function startDripping(address addr, uint256 multiplier) external virtual {
+    function startDripping(address addr, uint128 multiplier) external virtual {
         _startDripping(addr, multiplier);
     }
 
-    function stopDripping(address addr, uint256 multiplier) external virtual {
+    function stopDripping(address addr, uint128 multiplier) external virtual {
         _stopDripping(addr, multiplier);
     }
